Implement business logic for getting an activity board.

+++ authentication-and-member-microservice/src/main/java/nl/tudelft/sem/template/authmember/models/GetHoaModel.java
@@ -8,5 +8,5 @@ import lombok.Data;
 @Data
 public class GetHoaModel extends HoaModel {
     private String memberId;
-    private int hoaId;
+    private long hoaId;
 }
